pyproject.toml
[project]
name = "fastaulas"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "fastaulas", from = "src"}]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tpp.pytest.ini_opitions]
pythonpathh = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'fastapi dev src/fastaulas/app.py'
test = 'pytest --cache-clear -s -x --cov=src/fastaulas -vv'
post_test = 'coverage html'
pre_test = 'task lint'
format = 'ruff check . --fix && ruff format .'
lint = 'ruff check . && ruff check . --diff'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.4,<0.15.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]
